move private inference helpers into internal package
