+++ src/refiner/inference/client.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import httpx
 
-from refiner.inference._response import InferenceResponse
+from refiner.inference.internal.response import InferenceResponse
 from refiner.inference.providers.anthropic import _AnthropicEndpointClient
 from refiner.inference.providers.google import _GoogleEndpointClient
 from refiner.inference.providers.openai import (
