+++ src/refiner/inference/generate.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from collections.abc import Awaitable, Callable, Mapping
 from typing import Any, TypeAlias
 
-from refiner.inference._runtime import inference_map
-from refiner.inference._response import InferenceResponse
+from refiner.inference.internal.runtime import inference_map
+from refiner.inference.internal.response import InferenceResponse
 from refiner.inference.providers import OpenAIEndpointProvider, VLLMProvider
 from refiner.inference.providers.openai import _OpenAIEndpointClient
 from refiner.pipeline.data.row import Row
